+++ lib/elixir/test/elixir/binary/inspect_test.exs
@@ -86,12 +86,10 @@ defmodule Binary.Inspect.NumberTest do
     assert inspect(100) == "100"
   end
 
-  unless :erlang.system_info(:otp_release) < 'R16' do
-    test :float do
-      assert inspect(1.0) == "1.0"
-      assert inspect(1.0e10) == "10000000000.0"
-      assert inspect(1.0e-10) == "0.0000000001"
-    end
+  test :float do
+    assert inspect(1.0) == "1.0"
+    assert inspect(1.0e10) == "10000000000.0"
+    assert inspect(1.0e-10) == "0.0000000001"
   end
 end